+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 + Improve GSSAPI and Kebreros library loading
 + Added error messages to describe why PSOpenAD failed to find the implicit DC host
 + Change `-UseSSL` to `-UseTLS`
++ Always add `$` to the `-Identity` of `Get-OpenADComputer` and `Get-OpenADServiceAccount` when using a `sAMAccountName`
 
 ## v0.1.0-preview1 - 2022-02-1